Added changes for moving vastunwrap to openwrap
