+++ modules/pubmatic/openwrap/config/config.go
@@ -3,20 +3,22 @@ package config
 import (
 	"time"
 
+	unWrapCfg "git.pubmatic.com/vastunwrap/config"
 	"github.com/prebid/prebid-server/v2/modules/pubmatic/openwrap/metrics/stats"
 )
 
 // Config contains the values read from the config file at boot time
 type Config struct {
-	Server    Server
-	Database  Database
-	Cache     Cache
-	Timeout   Timeout
-	Tracker   Tracker
-	PixelView PixelView
-	Features  FeatureToggle
-	Log       Log
-	Stats     stats.Stats
+	Server        Server
+	Database      Database
+	Cache         Cache
+	Timeout       Timeout
+	Tracker       Tracker
+	PixelView     PixelView
+	Features      FeatureToggle
+	Log           Log
+	Stats         stats.Stats
+	VastUnwrapCfg unWrapCfg.VastUnWrapCfg
 }
 
 type Server struct {
@@ -80,6 +82,8 @@ type PixelView struct {
 }
 
 type FeatureToggle struct {
+	VASTUnwrapPecent      int
+	VASTUnwrapStatsPecent int
 }
 
 type Log struct { //Log Details
