+++ modules/pubmatic/openwrap/models/openwrap.go
@@ -92,13 +92,15 @@ type RequestCtx struct {
 	ReturnAllBidStatus     bool   // ReturnAllBidStatus stores the value of request.ext.prebid.returnallbidstatus
 	Sshb                   string //Sshb query param to identify that the request executed heder-bidding or not, sshb=1(executed HB(8001)), sshb=2(reverse proxy set from HB(8001->8000)), sshb=""(direct request(8000)).
 
-	DCName              string
-	CachePutMiss        int                                                          // to be used in case of CTV JSON endpoint/amp/inapp-ott-video endpoint
-	CurrencyConversion  func(from string, to string, value float64) (float64, error) `json:"-"`
-	MatchedImpression   map[string]int
-	CustomDimensions    map[string]CustomDimension
-	AmpVideoEnabled     bool //AmpVideoEnabled indicates whether to include a Video object in an AMP request.
-	IsTBFFeatureEnabled bool
+	DCName                 string
+	CachePutMiss           int                                                          // to be used in case of CTV JSON endpoint/amp/inapp-ott-video endpoint
+	CurrencyConversion     func(from string, to string, value float64) (float64, error) `json:"-"`
+	MatchedImpression      map[string]int
+	CustomDimensions       map[string]CustomDimension
+	AmpVideoEnabled        bool //AmpVideoEnabled indicates whether to include a Video object in an AMP request.
+	IsTBFFeatureEnabled    bool
+	VastUnwrapEnabled      bool
+	VastUnwrapStatsEnabled bool
 }
 
 type OwBid struct {
