+++ modules/pubmatic/vastunwrap/constant.go
@@ -1,26 +1,26 @@
-package vastunwrap
+ package vastunwrap
 
-const (
-	VastUnwrapEnabled   = "enableVastUnwrapper"
-	RequestContext      = "rctx"
-	UserAgent           = "User-Agent"
-	XUserIP             = "X-Forwarded-For"
-	XUserAgent          = "X-Device-User-Agent"
-	CreativeID          = "unwrap-ucrid"
-	PubID               = "pub_id"
-	ImpressionID        = "imr_id"
-	UnwrapCount         = "unwrap-count"
-	UnwrapStatus        = "unwrap-status"
-	Success             = "Success"
-	Failure             = "Failure"
-	Timeout             = "Timeout"
-	UnwrapStatusTimeout = "2"
-	UnwrapURL           = "http://localhost:8003/unwrap"
-	ContentType         = "Content-Type"
-	UnwrapTimeout       = "unwrap-timeout"
-	MediaTypeVideo      = "video"
-	ProfileId           = "profileID"
-	VersionId           = "versionID"
-	DisplayId           = "DisplayID"
-	Endpoint            = "Endpoint"
-)
+// const (
+// 	VastUnwrapEnabled   = "enableVastUnwrapper"
+// 	RequestContext      = "rctx"
+// 	UserAgent           = "User-Agent"
+// 	XUserIP             = "X-Forwarded-For"
+// 	XUserAgent          = "X-Device-User-Agent"
+// 	CreativeID          = "unwrap-ucrid"
+// 	PubID               = "pub_id"
+// 	ImpressionID        = "imr_id"
+// 	UnwrapCount         = "unwrap-count"
+// 	UnwrapStatus        = "unwrap-status"
+// 	Success             = "Success"
+// 	Failure             = "Failure"
+// 	Timeout             = "Timeout"
+// 	UnwrapStatusTimeout = "2"
+// 	UnwrapURL           = "http://localhost:8003/unwrap"
+// 	ContentType         = "Content-Type"
+// 	UnwrapTimeout       = "unwrap-timeout"
+// 	MediaTypeVideo      = "video"
+// 	ProfileId           = "profileID"
+// 	VersionId           = "versionID"
+// 	DisplayId           = "DisplayID"
+// 	Endpoint            = "Endpoint"
+// )
