+++ modules/pubmatic/vastunwrap/models/request.go
@@ -1,11 +1,11 @@
 package models
 
-type RequestCtx struct {
-	UA                          string
-	IP                          string
-	VastUnwrapEnabled           bool
-	PubID, ProfileID, DisplayID int
-	Endpoint                    string
-	VastUnwrapStatsEnabled      bool
-	Redirect                    bool
-}
+// type RequestCtx struct {
+// 	UA                          string
+// 	IP                          string
+// 	VastUnwrapEnabled           bool
+// 	PubID, ProfileID, DisplayID int
+// 	Endpoint                    string
+// 	VastUnwrapStatsEnabled      bool
+// 	Redirect                    bool
+// }
